service(django_14_10_24)/service(django_14_10_24)/requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
python-dateutil~=2.9.0.post0
django-environ~=0.11.2
django-tinymce~=4.1.0
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.1
django-environ==0.11.2
django-grappelli==4.0.1
django-tinymce==4.1.0
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
